Feat(score_edit): コンポーネントファイルに分けた
