+++ frontend/src/pages/score_edit/[uuid].tsx
@@ -5,20 +5,10 @@ import {
   Divider,
   Flex,
   Heading,
-  Input,
   Spacer,
   Text,
   VStack,
   Spinner,
-  Modal,
-  ModalOverlay,
-  ModalContent,
-  ModalHeader,
-  ModalBody,
-  ModalFooter,
-  ModalCloseButton,
-  FormControl,
-  FormLabel,
 } from '@chakra-ui/react';
 import { useDisclosure } from '@chakra-ui/hooks';
 import { SessionService } from 'services/SessionService';
@@ -33,6 +23,8 @@ import {
   ScoreUpdateRequestFactory,
   ScoreUpdateRequestObject,
 } from 'domains';
+import { AddScoreModal } from 'components/AddScoreModal';
+import { ScoreRow } from 'components/ScoreRow';
 
 const ScoreEdit = () => {
   // 良くない
@@ -286,137 +278,6 @@ const ScoreEdit = () => {
         </Container>
       )}
     </>
-  );
-};
-
-type ScoreRowProps = {
-  index: number;
-  gotScore: number;
-  maxScore: number;
-  onDelete: VoidFunction;
-  //onPressModal: VoidFunction;
-  onChangeMaxScore: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  onChangeGotScore: (e: React.ChangeEvent<HTMLInputElement>) => void;
-};
-
-const ScoreRow: React.FC<ScoreRowProps> = ({
-  index,
-  gotScore,
-  maxScore,
-  onDelete,
-  //onPressModal,
-  onChangeMaxScore,
-  onChangeGotScore,
-}) => {
-  return (
-    <Flex w='full' justifyContent='space-between'>
-      <Flex>
-        <Text color='whiteAlpha.900' lineHeight='full' w='2em'>
-          {index}.
-        </Text>
-        <Input
-          w='5rem'
-          bg='white'
-          color='black'
-          defaultValue={gotScore}
-          onChange={onChangeGotScore}
-          type='number'
-        ></Input>
-      </Flex>
-      <Flex>
-        <Input
-          w='5rem'
-          bg='white'
-          color='black'
-          defaultValue={maxScore}
-          onChange={onChangeMaxScore}
-          type='number'
-        ></Input>
-        <Button
-          bg='whiteAlpha.200'
-          color='whiteAlpha.800'
-          rounded='full'
-          ml={2}
-          mr={2}
-          onClick={onDelete}
-        >
-          削除
-        </Button>
-      </Flex>
-    </Flex>
-  );
-};
-
-type AddScoreModalProps = {
-  isOpen: boolean;
-  isDisabled: boolean;
-  onClose: () => void;
-  onClick: () => void;
-  onChangeMaxScore: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  onChangeGotScore: (e: React.ChangeEvent<HTMLInputElement>) => void;
-};
-
-const AddScoreModal: React.FC<AddScoreModalProps> = ({
-  isDisabled,
-  isOpen,
-  onClose,
-  onClick,
-  onChangeMaxScore,
-  onChangeGotScore,
-}) => {
-  return (
-    <Modal isOpen={isOpen} onClose={onClose} isCentered>
-      <ModalOverlay />
-      <ModalContent>
-        <ModalHeader>点数の追加</ModalHeader>
-        <ModalCloseButton />
-        <ModalBody pb={6}>
-          <Flex>
-            <FormControl id='gotScore'>
-              <FormLabel>獲得点数</FormLabel>
-              <Input
-                w='5rem'
-                bg='white'
-                color='black'
-                type='number'
-                onChange={onChangeGotScore}
-              />
-            </FormControl>
-            <FormControl id='maxScore'>
-              <FormLabel>満点</FormLabel>
-              <Input
-                w='5rem'
-                bg='white'
-                color='black'
-                type='number'
-                onChange={onChangeMaxScore}
-              />
-            </FormControl>
-          </Flex>
-        </ModalBody>
-        <ModalFooter>
-          <Button
-            pr={10}
-            pl={10}
-            mr={15}
-            bg='blue.400'
-            color='gray.800'
-            colorScheme={'blue'}
-            isDisabled={isDisabled}
-            onClick={onClick}
-            rounded='full'
-            _hover={{
-              bg: 'blue.300',
-            }}
-          >
-            追加
-          </Button>
-          <Button mr={10} onClick={onClose} rounded='full'>
-            キャンセル
-          </Button>
-        </ModalFooter>
-      </ModalContent>
-    </Modal>
   );
 };
